server/requirements.txt
pandas
torch
torchvision
pillow==8.2.0
tensorboardX
flask_cors
flask
